refactor plugin CI case and fork test case
Former-commit-id: f1ed1fac62dc5f46774df97abbb8ff511f9d5972
Former-commit-id: 0668fe025807beec13899112dcfc9dcd5a3350cd

+++ build/docker-compose-pre.sh
@@ -5,18 +5,22 @@ set -o pipefail
 #set -o verbose
 #set -o xtrace
 
+sedfix=""
+if [ "$(uname)" == "Darwin" ]; then
+    sedfix=".bak"
+fi
+
 OP="${1}"
 PROJ="${2}"
 DAPP="${3}"
 
-function run_dapp(){
+function run_dapp() {
     app=$1
-    if [ -d "${app}" ] && [ "${app}" != "system" ]; then
+    if [ -d "${app}" ]; then
         echo "============ run dapp=$app start ================="
-        rm -rf "${app}"-ci && mkdir -p "${app}"-ci && cp ./"${app}"/* ./"${app}"-ci && cp -n * ./"${app}"-ci/ && echo $?
+        rm -rf "${app}"-ci && mkdir -p "${app}"-ci && cp ./"${app}"/* ./"${app}"-ci && cp -n ./* ./"${app}"-ci/ && echo $?
         cd "${app}"-ci/ && pwd
-        if ! ./docker-compose.sh "${PROJ}" "${app}"
-        then
+        if ! ./docker-compose.sh "${PROJ}" "${app}"; then
             exit 1
         fi
         cd ..
@@ -25,31 +29,30 @@ function run_dapp(){
 
 }
 
-function down_dapp(){
+function down_dapp() {
     app=$1
-    if [ -d "${app}" ] && [ "${app}" != "system" ]; then
+    if [ -d "${app}" ] && [ "${app}" != "system" ] && [ -d "${app}-ci" ]; then
         cd "${app}"-ci/ && pwd
-        appfile="fork-test.sh"
-        if [ -e $appfile ]; then
-            echo "============ down dapp=$app start ================="
-            ./docker-compose-down.sh "${PROJ}" "${app}"
-            echo "============ down dapp=$app end ================="
-        fi
+
+        echo "============ down dapp=$app start ================="
+        ./docker-compose-down.sh "${PROJ}" "${app}"
+        echo "============ down dapp=$app end ================="
+
         cd .. && rm -rf ./"${app}"-ci
     fi
 }
 
-function forktest_dapp(){
+function forktest_dapp() {
     app=$1
-    if [ -d "${app}" ] ; then
+    if [ -d "${app}" ]; then
         echo "============ run dapp=$app start ================="
         rm -rf "${app}"-ci && mkdir -p "${app}"-ci && cp ./"${app}"/* ./"${app}"-ci && echo $?
-        cp -n * ./"${app}"-ci/ && echo $?
+        cp -n ./* ./"${app}"-ci/ && echo $?
         cd "${app}"-ci/ && pwd
+        sed -i $sedfix 's/^system_coins_file=.*/system_coins_file="..\/system\/coins\/fork-test.sh"/g' system-fork-test.sh
         appfile="fork-test.sh"
-        if [ -e $appfile ] || [ "${app}" == "system" ]; then
-            if ! ./system-fork-test.sh "${PROJ}" "${app}"
-            then
+        if [ -e $appfile ]; then
+            if ! ./system-fork-test.sh "${PROJ}" "${app}"; then
                 exit 1
             fi
         else
@@ -61,44 +64,81 @@ function forktest_dapp(){
 
 }
 
+function forktest_down() {
+    app=$1
+    if [ -d "${app}" ] && [ "${app}" != "system" ] && [ -d "${app}-ci" ]; then
+        cd "${app}"-ci/ && pwd
+        appfile="fork-test.sh"
+        if [ -e $appfile ]; then
+            echo "============ down dapp=$app start ================="
+            ./docker-compose-down.sh "${PROJ}" "${app}"
+            echo "============ down dapp=$app end ================="
+        fi
+        cd .. && rm -rf ./"${app}"-ci
+    fi
+}
+
 function main() {
-    if [ "${OP}" == "run" ];then
-        if [ "${DAPP}" == "all" ] ; then
+    if [ "${OP}" == "run" ]; then
+        if [ "${DAPP}" == "all" ]; then
             echo "============ run main start ================="
-    #        ./docker-compose.sh $PROJ
-    #        [ $? != 0 ] && exit 1
-    #        ./docker-compose-down.sh $PROJ
+            if ! ./docker-compose.sh "$PROJ"; then
+                exit 1
+            fi
+            ./docker-compose-down.sh "$PROJ"
             echo "============ run main end ================="
 
-            dir=$(ls -l ./ |awk '/^d/ {print $NF}')
-            for app in $dir
-            do
+            #dir=$(ls -l ./ |awk '/^d/ {print $NF}')
+            find . -maxdepth 1 -type d -name "*-ci" -exec rm -rf {} \;
+            dir=$(find . -maxdepth 1 -type d ! -name system ! -name . | sed 's/^\.\///')
+            for app in $dir; do
                 run_dapp "${app}"
                 down_dapp "${app}"
             done
-        elif [ -n "${DAPP}" ] ; then
+        elif [ -n "${DAPP}" ]; then
             run_dapp "${DAPP}"
         else
-           ./docker-compose.sh "${PROJ}"
+            ./docker-compose.sh "${PROJ}"
         fi
-    elif [ "${OP}" == "down" ];then
-        if [ -n "${DAPP}" ] ; then
+    elif [ "${OP}" == "down" ]; then
+        if [ "${DAPP}" == "all" ]; then
+            dir=$(find . -maxdepth 1 -type d ! -name system ! -name . | sed 's/^\.\///')
+            for app in $dir; do
+                down_dapp "${app}"
+            done
+            ./docker-compose-down.sh "${PROJ}"
+            remains=$(docker ps -a -q | awk '{print $1}')
+            echo $remains
+            for rems in $remains; do
+                docker rm "$rems" -f
+            done
+            docker ps -q
+        elif [ -n "${DAPP}" ]; then
             down_dapp "${DAPP}"
         else
             ./docker-compose-down.sh "${PROJ}"
         fi
-    elif [ "${OP}" == "forktest" ];then
-        if [ "${DAPP}" == "all" ] ; then
-            dir=$(ls -l ./ |awk '/^d/ {print $NF}')
-            for app in $dir
-            do
+    elif [ "${OP}" == "forktest" ]; then
+        if [ "${DAPP}" == "all" ]; then
+#            echo "============ run main start ================="
+#            if ! ./system-fork-test.sh "$PROJ"; then
+#                exit 1
+#            fi
+#            echo "============ down main test ================="
+#            ./docker-compose-down.sh "$PROJ"
+#            echo "============ run main end ================="
+            #  dir=$(ls -l ./ |awk '/^d/ {print $NF}')
+            ## remove all the *-ci folders
+            find . -maxdepth 1 -type d -name "*-ci" -exec rm -rf {} \;
+            dir=$(find . -maxdepth 1 -type d ! -name system ! -name . | sed 's/^\.\///')
+            for app in $dir; do
                 forktest_dapp "${app}"
-                down_dapp "${app}"
+                forktest_down "${app}"
             done
-        elif [ -n "${DAPP}" ] ; then
+        elif [ -n "${DAPP}" ]; then
             forktest_dapp "${DAPP}"
         else
-           forktest_dapp "system"
+            ./system-fork-test.sh "${PROJ}"
         fi
     fi
